+++ Controllers/StockController.cs
@@ -14,26 +14,28 @@ namespace api.Controllers
 {
     [Route("api/stock")]
     [ApiController]
-    public class StockController: ControllerBase
+    public class StockController : ControllerBase
     {
-        private readonly ApplicationDBContext _context;
+        // private readonly ApplicationDBContext _context;
         private readonly IStockRepository _stockRepo;
         public StockController(ApplicationDBContext context, IStockRepository stockRepo)
         {
-            _context = context;
+            // _context = context;
             _stockRepo = stockRepo;
         }
 
         [HttpGet]
-        public async Task<IActionResult> GetAll() {
+        public async Task<IActionResult> GetAll()
+        {
             var stocks = await _stockRepo.GetAllAsync();
             var stockDto = stocks.Select(s => s.ToStockDto());
 
             return Ok(stockDto);
         }
 
-        [HttpGet("{id}")]
-        public async Task<IActionResult> GetById([FromRoute] int id) {
+        [HttpGet("{id:int}")]
+        public async Task<IActionResult> GetById([FromRoute] int id)
+        {
             var stock = await _stockRepo.GetByIdAsync(id);
             if (stock == null)
             {
@@ -45,35 +47,45 @@ namespace api.Controllers
 
 
         [HttpPost]
-        public async Task<IActionResult> Create([FromBody] CreateStockRequestDto stockDto) {
+        public async Task<IActionResult> Create([FromBody] CreateStockRequestDto stockDto)
+        {
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
             var stockModel = stockDto.ToStockFromCreateDTO();
 
             await _stockRepo.CreateAsync(stockModel);
 
-            return CreatedAtAction(nameof (GetById), new {id = stockModel.Id}, stockModel.ToStockDto());
+            return CreatedAtAction(nameof(GetById), new { id = stockModel.Id }, stockModel.ToStockDto());
         }
 
         [HttpPut]
-        [Route("{id}")]
-        public async Task<IActionResult> Update([FromRoute] int id, UpdateStockRequestDto updateDto) {
+        [Route("{id:int}")]
+        public async Task<IActionResult> Update([FromRoute] int id, UpdateStockRequestDto updateDto)
+        {
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
+
             var stockModel = await _stockRepo.UpdateAsync(id, updateDto);
             if (stockModel == null)
             {
                 return NotFound();
             }
             return Ok(stockModel.ToStockDto());
-        } 
+        }
 
         [HttpDelete]
-        [Route("{id}")]
-        public async Task<IActionResult> Delete([FromRoute] int id) {
+        [Route("{id:int}")]
+        public async Task<IActionResult> Delete([FromRoute] int id)
+        {
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
             var stockModel = await _stockRepo.DeleteAsync(id);
             if (stockModel == null)
             {
                 return NotFound();
             }
             return NoContent();
-            
+
         }
 
     }
